refactor(self-custodial): drop crashlytics from helper, relocate fee-tier-options
