+++ app/screens/send-bitcoin-screen/hooks/fee-tiers.types.ts
@@ -1,0 +1,12 @@
+export const FeeTierOption = {
+  Fast: "fast",
+  Medium: "medium",
+  Slow: "slow",
+} as const
+
+export type FeeTierOption = (typeof FeeTierOption)[keyof typeof FeeTierOption]
+
+export type FeeTierInfo = {
+  feeSats: number
+  etaMinutes: number
+}
