+++ app/screens/send-bitcoin-screen/hooks/use-onchain-fee-tiers.ts
@@ -8,18 +8,7 @@ import {
 
 import { extractOnchainFees, prepareSend } from "@app/self-custodial/bridge"
 
-export const FeeTierOption = {
-  Fast: "fast",
-  Medium: "medium",
-  Slow: "slow",
-} as const
-
-export type FeeTierOption = (typeof FeeTierOption)[keyof typeof FeeTierOption]
-
-export type FeeTierInfo = {
-  feeSats: number
-  etaMinutes: number
-}
+import { FeeTierOption, type FeeTierInfo } from "./fee-tiers.types"
 
 export const ETA_MINUTES: Record<FeeTierOption, number> = {
   [FeeTierOption.Fast]: 10,
